Move Conclusion to be an inner class of Rule.
